requirements.txt
packaging>=21.3.0,<21.4.0
rich>=12.0.0,<12.1.0
pydantic>=1.9.0,<1.10.0
aiohttp>=3.8,<3.9
backports.cached-property>=1.0,<1.1
typing_extensions>=4.2,<4.3
